Rewrite Parts of MModding Config Module Float Elements becomes Double Elements

+++ mmodding-config/src/main/java/com/mmodding/library/config/api/content/ConfigContent.java
@@ -14,7 +14,7 @@ public interface ConfigContent {
 
 	int integer(String qualifier);
 
-	float floating(String qualifier);
+	double floating(String qualifier);
 
 	String string(String qualifier);
